:ok_hand: Move checkout to cart.js
