+++ app/handler/trades.js
@@ -2,7 +2,6 @@ const TradeOfferManager = require('steam-tradeoffer-manager');
 const Currencies = require('tf2-currencies');
 const pluralize = require('pluralize');
 const retry = require('retry');
-const SKU = require('tf2-sku');
 
 const log = require('lib/logger');
 const inventory = require('app/inventory');
@@ -12,13 +11,10 @@ const client = require('lib/client');
 const manager = require('lib/manager');
 const admin = require('app/admins');
 const groups = require('handler/groups');
-const schemaManager = require('lib/tf2-schema');
 
 const isAdmin = admin.isAdmin;
 const checkBanned = require('utils/isBanned');
 const communityLoginCallback = require('utils/communityLoginCallback');
-
-const cart = {};
 
 exports.getTradesWithPeople = function (steamIDs) {
     // Go through polldata data
@@ -76,288 +72,6 @@ exports.getActiveOffer = function (steamID) {
 
     return null;
 };
-
-exports.customOffer = function (partner, callback) {
-    const start = new Date().getTime();
-
-    if (cart[partner] === undefined) {
-        callback(null, 'Failed to send offer, your cart is empty');
-        return;
-    }
-
-    const alteredItems = { our: [], their: [] };
-    let alteredMessage;
-
-    // Check if we have all the items requested
-    for (const name in cart[partner].our) {
-        if (!Object.prototype.hasOwnProperty.call(cart[partner].our, name)) {
-            continue;
-        }
-
-        const amountCanTrade = inventory.findBySKU(cart[partner].our[name].sku, false).length;
-
-        if (cart[partner].our[name].amount > amountCanTrade) {
-            if (amountCanTrade === 0) {
-                delete cart[partner].our[name];
-            } else {
-                cart[partner].our[name].amount = amountCanTrade;
-            }
-            alteredItems.our.push(name);
-        }
-    }
-
-    if (Object.getOwnPropertyNames(cart[partner].our).length === 0 && Object.getOwnPropertyNames(cart[partner].their).length === 0) {
-        alteredMessage = createAlteredMessage(alteredItems, partner);
-        callback(null, alteredMessage);
-        return;
-    }
-
-    const offer = manager.createOffer(partner);
-
-    offer.data('partner', partner);
-
-    const itemsDict = { our: {}, their: {} };
-    const itemsDiff = {};
-
-    // Add our items to the trade
-    for (const name in cart[partner].our) {
-        if (!Object.prototype.hasOwnProperty.call(cart[partner].our, name)) {
-            continue;
-        }
-
-        const sku = cart[partner].our[name].sku;
-        const amount = cart[partner].our[name].amount;
-
-        const assetids = inventory.findBySKU(sku, false);
-
-        for (let i = 0; i < amount; i++) {
-            offer.addMyItem({
-                assetid: assetids[i],
-                appid: 440,
-                contextid: 2,
-                amount: 1
-            });
-        }
-
-        itemsDict.our[sku] = amount;
-        itemsDiff[sku] = amount*(-1);
-    }
-
-    if (Object.getOwnPropertyNames(cart[partner].their).length === 0) {
-        // We are not taking any items, don't request their inventory
-
-        alteredMessage = createAlteredMessage(alteredItems, partner);
-
-        if (Object.getOwnPropertyNames(cart[partner].our).length === 0) {
-            exports.removeFromCart(true, partner);
-            callback(null, alteredMessage);
-            return;
-        }
-
-        offer.data('dict', itemsDict);
-        offer.data('diff', itemsDiff);
-
-        offer.data('handleTimestamp', start);
-
-        offer.setMessage('Powered by TF2 Automatic');
-
-        client.chatMessage(partner, (alteredMessage || 'Please wait while I process your offer...'));
-
-        require('app/trade').sendOffer(offer, function (err) {
-            if (err) {
-                if (err.message.indexOf('We were unable to contact the game\'s item server') !== -1) {
-                    return callback(null, 'Team Fortress 2\'s item server may be down or Steam may be experiencing temporary connectivity issues');
-                } else if (err.message.indexOf('can only be sent to friends') != -1) {
-                    return callback(err);
-                } else if (err.message.indexOf('maximum number of items allowed in your Team Fortress 2 inventory') > -1) {
-                    return callback(null, 'I don\'t have space for more items in my inventory');
-                } else if (err.eresult !== undefined) {
-                    if (err.eresult == 10) {
-                        callback(null, 'An error occurred while sending your trade offer, this is most likely because I\'ve recently accepted a big offer');
-                    } else if (err.eresult == 15) {
-                        callback(null, 'I don\'t, or you don\'t, have space for more items');
-                    } else if (err.eresult == 16) {
-                        // This happens when Steam is already handling an offer (usually big offers), the offer should be made
-                        callback(null, 'An error occurred while sending your trade offer, this is most likely because I\'ve recently accepted a big offer');
-                    } else if (err.eresult == 20) {
-                        callback(null, 'Team Fortress 2\'s item server may be down or Steam may be experiencing temporary connectivity issues');
-                    } else {
-                        callback(null, 'An error occurred while sending the offer (' + TradeOfferManager.EResult[err.eresult] + ')');
-                    }
-                    return;
-                }
-            }
-
-            return callback(err);
-        });
-
-        return;
-    }
-
-    inventory.getDictionary(partner, false, function (err, theirDict) {
-        if (err) {
-            return callback(null, 'Failed to load inventories, Steam might be down');
-        }
-
-        for (const name in cart[partner].their) {
-            if (!Object.prototype.hasOwnProperty.call(cart[partner].their, name)) {
-                continue;
-            }
-
-            const sku = cart[partner].their[name].sku;
-
-            const assetids = (theirDict[sku] || []);
-            const amountCanTrade = assetids.length;
-
-            if (cart[partner].their[name].amount > amountCanTrade) {
-                if (amountCanTrade === 0) {
-                    delete cart[partner].their[name];
-                } else {
-                    cart[partner].their[name].amount = amountCanTrade;
-                }
-                alteredItems.their.push(name);
-            }
-
-            if (cart[partner].their[name] === undefined) {
-                continue;
-            }
-
-            const amount = cart[partner].their[name].amount;
-
-            for (let i = 0; i < amount; i++) {
-                offer.addTheirItem({
-                    assetid: assetids[i],
-                    appid: 440,
-                    contextid: 2,
-                    amount: 1
-                });
-            }
-
-            itemsDict.their[sku] = amount;
-            itemsDiff[sku] = (itemsDiff[sku] || 0) + amount;
-        }
-
-        alteredMessage = createAlteredMessage(alteredItems, partner);
-
-        if ((Object.getOwnPropertyNames(cart[partner].their).length === 0) && (Object.getOwnPropertyNames(cart[partner].our).length === 0)) {
-            exports.removeFromCart(true, partner);
-            callback(null, alteredMessage);
-            return;
-        }
-
-        offer.data('dict', itemsDict);
-        offer.data('diff', itemsDiff);
-
-        offer.data('handleTimestamp', start);
-
-        offer.setMessage('Powered by TF2 Automatic');
-
-        // TODO: Send the offer
-        // TODO: Create message if nothing is altered
-        client.chatMessage(partner, (alteredMessage || 'Please wait while I process your offer...'));
-
-        require('app/trade').sendOffer(offer, function (err) {
-            if (err) {
-                if (err.message.indexOf('We were unable to contact the game\'s item server') !== -1) {
-                    return callback(null, 'Team Fortress 2\'s item server may be down or Steam may be experiencing temporary connectivity issues');
-                } else if (err.message.indexOf('can only be sent to friends') != -1) {
-                    return callback(err);
-                } else if (err.message.indexOf('maximum number of items allowed in your Team Fortress 2 inventory') > -1) {
-                    return callback(null, 'I don\'t have space for more items in my inventory');
-                } else if (err.eresult !== undefined) {
-                    if (err.eresult == 10) {
-                        callback(null, 'An error occurred while sending your trade offer, this is most likely because I\'ve recently accepted a big offer');
-                    } else if (err.eresult == 15) {
-                        callback(null, 'I don\'t, or you don\'t, have space for more items');
-                    } else if (err.eresult == 16) {
-                        // This happens when Steam is already handling an offer (usually big offers), the offer should be made
-                        callback(null, 'An error occurred while sending your trade offer, this is most likely because I\'ve recently accepted a big offer');
-                    } else if (err.eresult == 20) {
-                        callback(null, 'Team Fortress 2\'s item server may be down or Steam may be experiencing temporary connectivity issues');
-                    } else {
-                        callback(null, 'An error occurred while sending the offer (' + TradeOfferManager.EResult[err.eresult] + ')');
-                    }
-                    return;
-                }
-            }
-
-            return callback(err);
-        });
-    });
-};
-
-function createAlteredMessage (alteredItems, steamID) {
-    const noneAvailable = { our: 'I don\'t have any', their: 'You don\'t have any' };
-    const someAvailable = { our: 'I only have', their: 'You only have' };
-
-    const none = { our: [], their: [] };
-    const some = { our: [], their: [] };
-
-    ['our', 'their'].forEach((whose) => {
-        alteredItems[whose].forEach((name) => {
-            if (cart[steamID][whose][name]) {
-                some[whose].push({ name, amount: cart[steamID][whose][name].amount });
-            } else {
-                none[whose].push(name);
-            }
-        });
-    });
-
-    ['our', 'their'].forEach((whose) => {
-        none[whose].forEach((name, i) => {
-            const last = (i === none[whose].length-1 && i > 0);
-
-            if (last) {
-                noneAvailable[whose] = noneAvailable[whose].slice(0, -1);
-                noneAvailable[whose] += ' or';
-            }
-
-            noneAvailable[whose] += ' ' + pluralize(name, 0);
-
-            if (!last && i > 0) {
-                noneAvailable[whose] += ',';
-            }
-        });
-    });
-
-    ['our', 'their'].forEach((whose) => {
-        some[whose].forEach((obj, i) => {
-            const name = obj.name;
-            const amount = obj.amount;
-
-            const last = (i === some[whose].length-1 && i > 0);
-
-            if (last) {
-                someAvailable[whose] = someAvailable[whose].slice(0, -1);
-                someAvailable[whose] += ' and';
-            }
-
-            someAvailable[whose] += ' ' + pluralize(name, amount, true);
-
-            if (!last && i > 0) {
-                someAvailable[whose] += ',';
-            }
-        });
-    });
-
-    let message = '';
-
-    ['our', 'their'].forEach((whose) => {
-        ['none', 'some'].forEach((which, i) => {
-            if (i) {
-                if (some[whose].length) {
-                    message += someAvailable[whose] + '\n';
-                }
-            } else {
-                if (none[whose].length) {
-                    message += noneAvailable[whose] + '\n';
-                }
-            }
-        });
-    });
-
-    return message;
-}
 
 exports.createOffer = function (details, callback) {
     const partner = details.steamid;
